cloud/review_portal/requirements.txt
flask>=2.0
gunicorn>=20.0
google-cloud-storage>=2.0
